k8s-requirements.txt
kubernetes>=24.0.0
google-cloud-storage>=2.10.0
prometheus-client>=0.17.0
pyyaml>=6.0
